Rename reduces/namespace to reducers/cluster.
