+++ dashboard/src/reducers/cluster.ts
