+++ dashboard/src/shared/types.ts
@@ -2,8 +2,8 @@ import * as jsonSchema from "json-schema";
 import { IOperatorsState } from "reducers/operators";
 import { IAuthState } from "../reducers/auth";
 import { IServiceCatalogState } from "../reducers/catalog";
+import { IClustersState } from "../reducers/cluster";
 import { IConfigState } from "../reducers/config";
-import { IClustersState } from "../reducers/namespace";
 import { IAppRepositoryState } from "../reducers/repos";
 import { hapi } from "./hapi/release";
 import { ResourceKind } from "./ResourceKinds";
